requirements.txt
aiogram == 2.10.1
SQLAlchemy == 1.3.19
python-dotenv==0.14.0
